requirements.txt
sqlalchemy
psycopg2-binary
yfinance
pandas
loguru
python-dotenvexchange_calendars
